+++ src/main/java/dev/solar/lotto/domain/PrizeMoney.java
@@ -1,4 +1,4 @@
-package dev.solar.lotto;
+package dev.solar.lotto.domain;
 
 import java.util.Arrays;
 import java.util.Optional;
